backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-genai
python-dotenv
pydantic
